+++ public/index.php
@@ -10,13 +10,7 @@ if (!is_readable($vendor)) {
     exit('Site down for maintenance');
 }
 require_once($vendor);
-if (!class_exists('\Attogram\Router\Router')
-    || !class_exists('\Attogram\Justrefs\Web')
-    || !class_exists('\Attogram\Justrefs\Filesystem')
-    || !class_exists('\Attogram\Justrefs\Mediawiki')
-) {
-    exit('Site down for maintenance.');
-}
-$jr = new \Attogram\Justrefs\Web();
+
+$jr = new \Attogram\Justrefs\JustRefs();
 $jr->verbose = $verbose;
 $jr->route();
