+++ public/style.css
@@ -47,7 +47,7 @@ a.missing {
 }
 .head {
   background-color:lightgrey;
-  padding:10px 10px 10px 10px;
+  padding:8px 8px 8px 8px;
 }
 .body {
   background-color:white;
