+++ src/Template.php
@@ -3,7 +3,6 @@
  * Just Refs - https://github.com/attogram/justrefs
  *
  * Template Class
- *
  */
 declare(strict_types = 1);
 
@@ -11,8 +10,9 @@ namespace Attogram\Justrefs;
 
 use function is_array;
 use function is_readable;
+use function is_string;
 
-class Template
+class Template extends Base
 {
     /**
      * @param string $templateDirectory - path to template directory, with trailing slash
@@ -25,36 +25,47 @@ class Template
     private $vars = [];
 
     /**
-     * @param array $vars
-     */
-    public function setVars($vars)
-    {
-        if (!is_array($vars)) {
-            $vars = [];
-        }
-        $this->vars = $vars;
-    }
-
-    /**
+     * include a template
      * @param string $name
      * @return bool
      */
-    public function include($name)
+    public function include(string $name): bool
     {
         $template = $this->templateDirectory . $name . '.php';
         if (!is_readable($template)) {
             return false;
         }
-        include($template);
+        try {
+            include($template);
+        } catch (Exception $exception) {
+            return false;
+        }
+
+        return true;
+    }
+    
+    /**
+     * set a single var
+     * @param string $name
+     * @param string|mixed $value
+     */
+    public function set(string $name, $value): bool
+    {
+        if (!is_string($name)) {
+            return false;
+        }
+        $this->vars[$name] = $value;
+        $this->verbose("set: $name = " . print_r($value, true));
 
         return true;
     }
 
     /**
+     * get a var
      * @param string $name
      * @return string|mixed
      */
-    public function var($name)
+    public function get($name)
     {
         if (!isset($this->vars[$name])) {
             return '';
