+++ templates/footer.php
@@ -5,9 +5,9 @@
  * 
  * @uses $this - Attogram\Justrefs\Template
  */
-?>
-</div><footer>
-<a href="<?= $this->var('home') ?>"><?= $this->var('title') ?></a> 
-- <a href="<?= $this->var('home') ?>about/">About</a>
-<br /><small>page generated in <?= $this->endTimer('page') ?> seconds</small>
-</footer></body></html>
+?><footer>
+  <b><a href="<?= $this->get('home') ?>"><?= $this->get('name') ?></a></b>
+  <small>v<?= $this->get('version') ?></small>
+  - <small>page generated in <?= $this->endTimer('page') ?> seconds</small>
+</footer>
+</body></html>
